requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
gunicorn==22.0.0
pydantic==2.7.0
pydantic-settings==2.2.1
python-multipart==0.0.9
aiofiles==23.2.1
vaderSentiment==3.3.2
nltk==3.8.1
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.4
anthropic==0.26.1
pinecone-client==3.2.2
sentence-transformers==2.7.0
torch==2.3.0
transformers==4.41.0
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
numpy==1.26.4
gTTS==2.5.1
supabase==2.4.6
python-jose==3.3.0